Allow Variables in the (legacy) THNN bindings. (#4723)
The legacy NN bindings currently operate only on Tensors. We are slowly
replacing all uses of Tensor with Variable in Python code so that
there
will only be one user-visible class. This changes the NN bindings
accessed through type2backend to accept either Tensors or Variables.

This does not affect the NN bindings that go through ATen.